pyproject.toml
[tool.poetry]
name = "fmc-db"
version = "0.1.0"
description = ""
authors = ["jacoacoacob <jacoacoacob@protonmail.com>"]
readme = "README.md"
packages = [
  { include = "scripts" }
]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
psycopg2 = "^2.9.5"

[tool.poetry.scripts]
"cli" = "scripts.cli:cli.run"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
